+++ lisp/eab-workgroups2.el
@@ -31,10 +31,13 @@
   (let* ((true-path (file-truename path))
 	 (nondir (file-name-nondirectory path))
 	 (name (concat ":" nondir ":")))
-    (wg-create-workgroup name)
-    (dired true-path)
-    (eab/wg-update-workgroup "dflt")
-    (wg-save-session t)))
+    (if (not (wg-get-workgroup name 't))
+	(progn
+	  (wg-create-workgroup name)
+	  (dired true-path)
+	  (eab/wg-update-workgroup "dflt")
+	  (wg-save-session t)))
+    (eab/wg-add-workgroup-to-history (wg-workgroup-uid (wg-get-workgroup name 't)))))
 
 ;; TODO можно использовать `gr list` вместо wg/*: все-равно вручную
 ;; пополняю оба эти списка хотя, симлинки требуют меньше зависимостей